Change the display brightness on button presses.
Use timer 2 to generate a PWM signal on the LCD brightness pin, and
allow five brightness levels.  The lowest level has the back light
off, and the next higher level is the lowest we can achieve with the
back light still on, for viewing in pitch black conditions.  Move
between the brightness levels on the up and down buttons.

Signed-off-by: Russell Steicke <russells@adelie.cx>