Allow the user to set the time.
The user can now press the select button to start setting the time, up
or down to set the hours, minutes, or seconds, and select also to move
between hours, minutes, or seconds.  The time is only changed if any
of the three were actually modified.  The current time is monitored in
the background while the user steps through the hours, minutes, and
seconds, and if none are modified, the clock returns to the current
time at the end.  If any are modified, then that new value is used.

This adds a group of time setting states to the DClock active object.
There are also new event types to indicate what part of the time
setting process we are doing or entering.  A fair bit of this is done
by DClock sending signals to itself, which is really communication
between parent and child states.

Signed-off-by: Russell Steicke <russells@adelie.cx>