Time out while waiting for the time to be set.
We present a word that says what the user is setting ("Hours",
"Minutes", or "Seconds") and a line of five dots.  The line of dots
decreases by one every two (real) seconds, and when it gets to zero,
we go back to displaying the time without changing it.  The time will
appear to jump ahead, but that's only because it has continue to be
counted in the background while the time setting display was static.

If we get a timeout, the time is not set even if it was edited before
the timeout.

Signed-off-by: Russell Steicke <russells@adelie.cx>