Make better macros for event queue checking.
The macro that we used for checking the event queue was not ideal, and
badly named.  Create two macros to do the check and the post, one for
ISR code and one for non-ISR code, so we don't have to remember to
call the check before posting.  Use them everywhere an event is
posted.

Signed-off-by: Russell Steicke <russells@adelie.cx>