Swap the set time, and move dots to the bottom line.
When setting the time, put the words on the top line, the time on the
bottom line, and the dots on the bottom line.  This makes a bit more
sense in the context of a menu item, where the instructions are above
the thing we're doing.  Putting the timeout dots on the line with the
time (where there's plenty of room for them) means that the words can
fill up the whole 16 positions on the other line.

Signed-off-by: Russell Steicke <russells@adelie.cx>