With the LCD off, the select button temporarily increases brightness.
The LCD brightness has an off state, where no light is emitted at all.
In that state it makes no sense to try to set anything, so the select
key should not be used for setting the time then.  It can now be used
to temporarily increase the brightness so the user can read the time.

The brightness increases in increments as the key is held down for a
while.

Signed-off-by: Russell Steicke <russells@adelie.cx>