Add a simple alarm.
This will only alarm at a single time, 1.17.49, which is four decimal
seconds after our default start time.  It flashes the display between
two brightnesses and flashes "Alarm!" on the bottom line until the
select button is pressed.

All button presses now go through the alarm before they are forwarded
to the clock.  This is so the alarm can block button presses while it
is busy flashing and alarming.

If we are setting the time when the alarm fires, nothing happens
because TICK_DECIMAL_SIGNAL events are being used only to increment
the seconds at that point, and telling the alarm about the current
time is done by a parent state that does not receive that event.

Signed-off-by: Russell Steicke <russells@adelie.cx>