Include the QP-nano files in the source.
This copies the relevant QP-nano source files into the project, so we
can make changes to QP within the project.  We leave the qp-nano
symlink in place, but include that in the project as well, so we can
replace the QP-nano version if required.

Because the free version of QP-nano is GPL, this application is
probably also GPL now.  OK.

Signed-off-by: Russell Steicke <russells@adelie.cx>