Allow a QP-nano timer to send a specified signal at timeout.
So that we can tell timeout signals apart when they are generated by
different states, modify QP-nano to allow specification of a signal to
be generated at timer expiry.

We now use different signals in each of the states that set the hours,
minutes, and seconds, as there was a race condition in there that
could be triggered if the user pressed a button just as the timer was
expiring.  The next setting state or the parent state could have
gotten a stale timeout signal, but can't now because all the timeout
signals are different, and the "wrong" timeout signal from the
previous state will be ignored.

Idea from Harry McNally.

Signed-off-by: Russell Steicke <russells@adelie.cx>