When the backlight is off, turn off the LCD completely.
The intended reasons for turning the backlight brightness to its
lowest value (which is off) are to allow the user to sleep without the
light, and to preserve the LCD life.  But the backlight makes no
difference to the LCD life, as the LCD is still active with the
backlight off - it doesn't generate its own light but is still active
and visible with reflected light.  So turn off the LCD completely when
the backlight is off.

Signed-off-by: Russell Steicke <russells@adelie.cx>