Increase the serial rate to get more info out.
At 115200 bps serial rate and 16MHz system clock, we have 1388 clock
cycles per transmitted byte.  There are only 50 instructions in the
serial interrupt handler, so there is plenty of time available for the
higher rate.

At 9600 bps, we would sometimes lose information when an assertion
triggered, as the data would not have been transmitted before the
assertion handler was called.  This should be less of a problem with a
higher sreial rate.

Signed-off-by: Russell Steicke <russells@adelie.cx>