Disable the watchdog at startup.
When the watchdog timer is enabled at startup, it defaults to the
shortest timeout, 16ms.  If we wait longer than that to change the
timeout, we'll get a reset.  This could happen if we wait for serial
output to drain.  So disable the watchdog timer early, and allow it to
be reenabled when everything else is set up.

Also, drain the serial output just before the watchdog timer is
reenabled, so we don't trigger this problem later.

Signed-off-by: Russell Steicke <russells@adelie.cx>