Read the time from an I2C clock.
Written for the DS3232, as on the Freetronics RTC Module.

This only reads the time once, at startup, and never writes the time
back to the RTC.  There should probably be an active object whose only
function is to keep the time, separating that out from the display and
menu handling.

Signed-off-by: Russell Steicke <russells@adelie.cx>