Add the normal time mode.
The clock is now a normal clock as well as a decimal clock.  The
"secret" button sequence causes the mode to change.

We now carry the time around in two different formats at different
times.  One is the decimal time, which is a 32 bit unsigned integer
seconds count since the beginning of the day, and the other is hours
minutes and seconds in a struct whose size is 32 bits.  When the time
or alarm time is set, that time is stored once as both a normal time
and a decimal time, to avoid rounding errors caused by converting back
and forth.  Where the time mode is important, get_time_mode() is used
to distinguish them.

Tracking the normal time is done with a 1Hz interrupt from the RTC,
whereas tracking the decimal time is still done with a timer run from
the CPU clock, so the two times will slowly diverge.  Synchronising
them at the 108/125 boundaries has not been done, yet.

Signed-off-by: Russell Steicke <russells@adelie.cx>