Show the alarm state on the bottom line of the display.
The text showing the alarm state can be blank (no alarm set), "A"
(alarm set), "Zzz.." (snoozing), or "** ALARM **" (alarm active).  The
text scrolls across the display each minute for alarm active or each
second for the others.

We implement this with an ordered bitmap of statuses, and only the
highest priority active status is shown.

Signed-off-by: Russell Steicke <russells@adelie.cx>