Make noise and light when the alarm happens.
We now sound a buzzer and flash two RGB LEDs when the alarm runs.  The
colours on the LEDs are selected randomly.  The buzzer has an
alternating chirping sound.

Also, when the user holds the select button for a while to turn off
the alarm (rather than snoozing with a short press) there is a short
beep as the alarm decides to turn off.  This gives confirmation that
the alarm has turned off, and the user doesn't have to guess how long
to hold the button.

Signed-off-by: Russell Steicke <russells@adelie.cx>